Fixed the URL structure of the Poole
Changes the URL structure of the Poole so that there will not be double
slashes '//' in the URL while Jekyll build the site. Because making
`relative_permalinks: true` do not remove the leading slashes in the
post URL. @parkr said, "In Jekyll `relative_permalinks: true` means that
when I have about/index.html, I can set permalink: / in that file and
the permalink will go to /about/ rather than /about/index.html.
Basically that the permalink of a file is relative to its source
directory. URL's will always contain the preceding /."