Change Atom feed id to {{ site.url }}{{ site.baseurl }}
If the site is hosted on sub directory (eg http://example.com/blog) then
using only the site.url (http://example.com) may conflict with the id of
the feed of the site hosted on the domain.

It's better to use {{ site.url }}{{ site.baseurl }}
(http://example.com/blog) to avoid conflicts.